requirements.txt
Django~=2.0.6
# No newline at the EOF
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
colorama==0.4.1
coverage==4.5.4
decorator==4.4.0
Django==2.2.4
django-crispy-forms==1.7.2
django-extensions==2.2.1
django-extra-views==0.12.0
Faker==0.9.1
graphviz==0.13
importlib-metadata==0.19
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
mixer==6.1.3
more-itertools==7.2.0
packaging==19.1
parso==0.5.1
pickleshare==0.7.5
pluggy==0.12.0
prompt-toolkit==2.0.9
py==1.8.0
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.4.2
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pyparsing==2.4.2
pytest==5.1.1
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-html==1.22.0
pytest-ipdb==0.1
pytest-metadata==1.8.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.5.2
django-ckeditor==5.7.1